Redesigned password edit texts
An icon has been added, when you click on it, the entered password is displayed or disappears. Two extension functions for Edittext and ImageView have also been added. The first function checks the current Input Type and replaces it with another one, the second function checks in which position ImageView is pressed and replaces the icon color accordingly.